update BasePiece and Bishop, by namespace
